pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "guardspine-connector-template"
version = "0.1.0"
description = "Template for building GuardSpine connectors"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "GuardSpine Contributors" }
]
keywords = ["guardspine", "connector", "evidence", "governance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "httpx>=0.25",
    "pydantic>=2.0",
    "watchfiles>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]
github = [
    "pygithub>=2.0",
]
sharepoint = [
    "msal>=1.24",
    "msgraph-core>=1.0",
]
google = [
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
]

[project.scripts]
guardspine-connector = "connector.cli:main"

[project.urls]
Homepage = "https://github.com/guardspine/guardspine-connector-template"
Documentation = "https://guardspine.io/docs/connectors"
